Add Alias operation to redirect to a resource Retrieve endpoint
This convention allows for "vanity" URLs that alias UUID-based resources using a human-
readable name (as long as the name is valid as part of the URL path).

A mapping for an Alias operation supports a `GET` based request with a path parameter
of the form `foo_name`. The mapping function is expected to return a valid resource
(or raise `NotFound`); the convention will issue a redirect to the target resource.